+++ docs/pages/components/modal/Modal.vue
@@ -58,8 +58,8 @@
                 this.$modal.open(props)
 
                 // From outside Vue instance
-                import { ModalProgrammatic } from 'buefy/dist/components/modal'
-                ModalProgrammatic.open(props)`
+                import { ModalProgrammatic as Modal } from 'buefy'
+                Modal.open(props)`
             }
         }
     }