+++ src/index.js
@@ -25,4 +25,12 @@ use(Buefy)
 
 export default Buefy
 
+// export all components as vue plugin
 export * from './components'
+// export programmatic component
+export { DialogProgrammatic } from './components/dialog'
+export { LoadingProgrammatic } from './components/loading'
+export { ModalProgrammatic } from './components/modal'
+export { NotificationProgrammatic } from './components/notification'
+export { SnackbarProgrammatic } from './components/snackbar'
+export { ToastProgrammatic } from './components/toast'